+++ jdaviz/core/loaders/tests/test_loaders.py
@@ -10,6 +10,7 @@ from astropy.table import Table
 from astropy.io import fits
 from astropy.wcs import WCS
 from astroquery.mast import Mast, MastMissions
+from astropy.utils.data import download_file
 from gwcs import WCS as GWCS
 from specutils import SpectralRegion, Spectrum
 
@@ -1394,3 +1395,16 @@ def test_load_cube_no_dq_in_viewer(deconfigged_helper):
     datasets = deconfigged_helper.datasets
     assert len(datasets) == 3
     assert '3D Spectrum [DQ]' in datasets
+
+
+@pytest.mark.remote_data
+def test_niriss_image_selection(deconfigged_helper):
+    fpath = download_file('https://stsci.box.com/shared/static/w5ycpuue07mpltam3apn9fsi5ajw2euw.fits')  # noqa
+    ldr = deconfigged_helper.loaders['file']
+    # We don't actually have to load the file, since the transposed warning triggers on parsing
+    ldr.filepath = fpath
+
+    logger = deconfigged_helper.plugins['Logger']
+    for msg in logger.history:
+        if msg['text'][0:10] == 'Transposed':
+            raise ValueError('NIRISS images should not be treated as 2D spectra.')